Optimizing float conversion on CPU side.

+++ include/cloudVolume.h
@@ -8,7 +8,7 @@ struct CloudVolume
     CloudVolume() {}
     CloudVolume(glm::vec3 minimum, glm::vec3 maximum): minimum(minimum), maximum(maximum) {}
 
-    double resolution;
+    float resolution;
     glm::vec3 minimum;
     glm::vec3 maximum;
 };
